backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
requests==2.32.5
pandas==2.3.2
cryptography==45.0.7
redis==6.4.0
